refactor: rename source files to kebab-case, setAttempt to attemptCount
unicorn/filename-case flagged the 4 camelCase source files; renamed
with git mv and updated every import site:
  onkyoReceiver.ts -> onkyo-receiver.ts
  onkyoPlatform.ts -> onkyo-platform.ts
  receiverConfig.ts -> receiver-config.ts
  receiverInputConfig.ts -> receiver-input-config.ts

unicorn/no-non-function-verb-prefix flagged the setAttempt field (a
plain counter, not a function, despite the set* name) - renamed to
attemptCount at all 11 sites.

(The interface-to-type conversions originally planned for this step
were already handled as a side effect of the previous xo --fix
commit.)
