+++ src/eiscp/config.ts
@@ -1,5 +1,5 @@
 import type { PlatformConfig } from 'homebridge';
-import type { ReceiverConfig } from '../receiverConfig.js';
+import type { ReceiverConfig } from '../receiver-config.js';
 
 export interface Config extends PlatformConfig {
   receivers: ReceiverConfig[];
