+++ src/receiver-input-config.ts
